+++ src/utils/index.ts
@@ -15,6 +15,7 @@ export { VersionManager } from './versionManager';
 export { ZhipuWizard } from './zhipuWizard';
 export { MiniMaxWizard } from './minimaxWizard';
 export { MoonshotWizard } from './moonshotWizard';
+export { DashscopeWizard } from './dashscopeWizard';
 export { JsonSchemaProvider } from './jsonSchemaProvider';
 export { RetryManager } from './retryManager';
 export { ModelInfoCache } from './modelInfoCache';